fix: correct float conversion in visualize_sensitivity and update log

+++ Code/visualize_sensitivity.py
@@ -134,7 +134,7 @@ def main():
     # Calculate Phase Speed
     pspeed_array = np.array([(value.real / k_cal[:, None]) * 50 for value in disp_roots.values()])
     
-    labels = list([key.split("=")[-1].astype(float) for key in disp_roots.keys()])
+    labels = list([float(key.split("=")[-1]) for key in disp_roots.keys()])
     labels_argsort = np.argsort(np.array(labels))
 
     x_ticks = k_dis[demo_kidx]
